refactor : move modal state to separate hook
